+++ packages/components/anchor/utils.ts
@@ -1,3 +1,5 @@
+import Clipboard from 'clipboard';
+
 export const ANCHOR_SHARP_REGEXP = /#(\S+)$/;
 
 export type ANCHOR_CONTAINER = HTMLElement | Window;
@@ -9,4 +11,17 @@ export function getOffsetTop(target: HTMLElement, container: ANCHOR_CONTAINER): 
     return top - document.documentElement.clientTop;
   }
   return top - (container as HTMLElement).getBoundingClientRect().top;
+}
+
+// TODO：仅为这一个而引入一个 clipboard，可优化
+export function copyText(text: string) {
+  const div = document.createElement('div');
+  const clip = new Clipboard(div, {
+    text() {
+      return text;
+    },
+  });
+  div.click();
+  clip.destroy();
+  div.remove();
 }
